Configuration change
Getting No handler bound for FailedTask error. I think there is need for
binding the parameter to LineCounter at Driver

+++ src/main/java/edu/snu/cms/reef/tutorial/LineCountingTask.java
@@ -53,7 +53,7 @@ public class LineCountingTask implements Task {
   private String [] attrStr;
   @Inject
   public LineCountingTask(final DataSet<LongWritable, Text> dataSet,
-		  				  final @Parameter(LearnRate.class) int learnRate,
+		  				  final @Parameter(LearnRate.class) double learnRate,
 		  				  final @Parameter(NumParam.class) int numParam,
 		  				  final @Parameter(TargetParam.class) int targetParam) {
     this.dataSet = dataSet;
